requirements.txt
python-telegram-bot
pytesseract